# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ASKO API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/chat/conversations": {
      "post": {
        "operationId": "ChatController_createConversation",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "operationId": "ChatController_listConversations",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedConversationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversations/{id}": {
      "get": {
        "operationId": "ChatController_getConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "delete": {
        "operationId": "ChatController_deleteConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversations/{id}/participants": {
      "get": {
        "operationId": "ChatController_listParticipants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "operationId": "ChatController_addParticipant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversations/{id}/participants/{userId}": {
      "delete": {
        "operationId": "ChatController_removeParticipant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/conversations/{id}/messages": {
      "post": {
        "operationId": "ChatController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "operationId": "ChatController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "beforeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/messages/{id}": {
      "put": {
        "operationId": "ChatController_updateMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      },
      "delete": {
        "operationId": "ChatController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/chat/presence/{userId}": {
      "get": {
        "operationId": "ChatController_getPresence",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/presence/bulk": {
      "post": {
        "operationId": "ChatController_getBulkPresence",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPresenceResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/search-users": {
      "get": {
        "operationId": "ChatController_searchUsers",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "ConversationType": {
        "type": "string",
        "enum": [
          "direct",
          "group"
        ]
      },
      "ParticipantRole": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "member"
        ]
      },
      "ParticipantRecordDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantRole"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "lastReadMessageId": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string"
          }
        },
        "required": [
          "role",
          "id",
          "userId",
          "conversationId",
          "joinedAt"
        ]
      },
      "MessageType": {
        "type": "string",
        "enum": [
          "text",
          "image",
          "video",
          "document",
          "system"
        ]
      },
      "MessageStatus": {
        "type": "string",
        "enum": [
          "sending",
          "delivered",
          "seen"
        ]
      },
      "ChatMessageRecordDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageType"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "attachmentJson": {
            "type": "string"
          },
          "isEdited": {
            "type": "boolean"
          },
          "deliveredAt": {
            "type": "string"
          },
          "readAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "conversationId",
          "senderId",
          "isEdited",
          "createdAt",
          "updatedAt"
        ]
      },
      "ConversationRecordDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ConversationType"
              }
            ]
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantRecordDto"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "creatorId": {
            "type": "string"
          },
          "lastMessage": {
            "$ref": "#/components/schemas/ChatMessageRecordDto"
          },
          "unreadCount": {
            "type": "number"
          },
          "avatarUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "participants",
          "id",
          "creatorId",
          "unreadCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "ConversationResponseDto": {
        "type": "object",
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationRecordDto"
          }
        },
        "required": [
          "conversation"
        ]
      },
      "PaginatedConversationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationRecordDto"
            }
          },
          "overallCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "overallCount",
          "page",
          "limit"
        ]
      },
      "ParticipantListResponseDto": {
        "type": "object",
        "properties": {
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantRecordDto"
            }
          }
        },
        "required": [
          "participants"
        ]
      },
      "ChatMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessageRecordDto"
          }
        },
        "required": [
          "message"
        ]
      },
      "PaginatedMessagesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageRecordDto"
            }
          },
          "overallCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "overallCount",
          "page",
          "limit"
        ]
      },
...
      "PresenceStatus": {
        "type": "string",
        "enum": [
          "online",
          "offline"
        ]
      },
      "UserActivity": {
        "type": "string",
        "enum": [
          "idle",
          "typing",
          "uploading_image",
          "uploading_video",
          "uploading_document"
        ]
      },
      "PresenceRecordDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PresenceStatus"
              }
            ]
          },
          "activity": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserActivity"
              }
            ]
          },
          "userId": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "activity",
          "userId",
          "lastSeenAt"
        ]
      },
      "PresenceResponseDto": {
        "type": "object",
        "properties": {
          "presence": {
            "$ref": "#/components/schemas/PresenceRecordDto"
          }
        },
        "required": [
          "presence"
        ]
      },
      "BulkPresenceResponseDto": {
        "type": "object",
        "properties": {
          "presences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PresenceRecordDto"
            }
          }
        },
        "required": [
          "presences"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    }
  }
}